Exeption

+++ src/main/java/com/luizmrd/services/UserService.java
@@ -2,9 +2,11 @@ package com.luizmrd.services;
 
 import com.luizmrd.entities.User;
 import com.luizmrd.repositories.UserRepository;
+import com.luizmrd.resources.exeptions.DataBaseException;
 import com.luizmrd.services.exeptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,17 +31,22 @@ public class UserService {
     }
 
     public void delete(Long id){
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
         repository.deleteById(id);
     }
 
+
     public User update(Long id, User obj){
         User entity = repository.getReferenceById(id);
-        updadeData(entity,obj);
+        updateData(entity,obj);
         return repository.save(entity);
 
     }
 
-    private void updadeData(User entity, User obj) {
+    private void updateData(User entity, User obj) {
         entity.setEmail(obj.getEmail());
         entity.setName(obj.getName());
         entity.setPhone(obj.getPhone());
